cmd/snapshots: Add option to limit snapshots list
This patch adds a `--limit` option to limit snapshots list to the n
last snapshots. It is very similar to the `--last` one but does not
limit to one entry.

Output example:

    $ restic snapshots --limit 2
    repository 0d3eb989 opened successfully, password is correct
    ID        Time                 Host        Tags        Paths
    ------------------------------------------------------------
    5a33bdcc  2020-12-14 12:30:00  local                   /home
    73887d8e  2020-12-15 12:30:00  local                   /home
    ------------------------------------------------------------
    2 snapshots

To keep compatibility with previous versions a new function has been
added: `FilterSnapshotsWithLimit`. This function that filters a list
of snapshots to only return the N last entries for each hostname and
path. `FilterLastSnapshots` calls `FilterSnapshotsWithLimit(list, 1)`.

Signed-off-by: Sébastien Gross <seb•ɑƬ•chezwam•ɖɵʈ•org>